+++ webpack.release.js
@@ -1,7 +1,6 @@
-const webpack = require('webpack');
 const merge = require('webpack-merge');
 const base = require('./webpack.base.js');
-const UglifyJsPlugin = require('uglifyjs-webpack-plugin')
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 
 module.exports = merge(base, {
     mode: 'production',
@@ -10,9 +9,11 @@ module.exports = merge(base, {
     },
     optimization: {
         noEmitOnErrors: true,
-        minimizer: [new UglifyJsPlugin({
-            parallel: true,
-            sourceMap: false
-        })]
+        minimizer: [
+            new UglifyJsPlugin({
+                parallel: true,
+                sourceMap: false
+            })
+        ]
     }
 });