+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+
+- Change dev stack to Babel, Jest and ESLint.
+
 ## [1.1.0 - 2018-06-29](https://github.com/alonrbar/peppermint-di/tree/v1.1.0)
 
 ### Added