+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - **BREAKING**: `ResolveOptions.params` is now a `Map` instead of a plain object.
 - Change dev stack to Babel, Jest and ESLint.
+- Removed `lodash.defaultsdeep` dependency.
 
 ## [1.1.0 - 2018-06-29](https://github.com/alonrbar/peppermint-di/tree/v1.1.0)